Completed refator chicken health record flow
