+++ db/health_statusDB.js
@@ -1,0 +1,8 @@
+import pool from './database.js';
+
+// Get chicken health status
+export async function getHealthStatus () {
+  const [result] = await pool.query(`
+  SELECT * FROM health_status`);
+  return result;
+}
