+++ routes/chickenTransferRoutes.js
@@ -1,0 +1,15 @@
+import express from 'express';
+import * as ChickenTransferCtrl from '../controller/chickenTransferCtrl.js';
+
+const router = express.Router();
+
+// View Chicken Record Page
+router.get('/view', ChickenTransferCtrl.getChickenRecordPage);
+
+// View Chicken Transfer Form
+router.get('/create', ChickenTransferCtrl.getChickenTransferForm);
+
+// Submit Chicken Transfer Form
+router.post('/submit', ChickenTransferCtrl.submitChickenTransferForm);
+
+export default router;
