+++ routes/routes.js
@@ -1,5 +1,6 @@
 import CoopRoutes from './coopRoutes.js';
-import ChickenRoutes from './chickenRoutes.js';
+import ChickenTransferRoutes from './chickenTransferRoutes.js';
+import ChickenHealthRoutes from './chickenHealthRoutes.js';
 import BrooderRoutes from './brooderRoutes.js';
 import IncubatorRoutes from './incubatorRoutes.js';
 
@@ -7,7 +8,8 @@ import IncubatorRoutes from './incubatorRoutes.js';
 
 export {
   CoopRoutes,
-  ChickenRoutes,
+  ChickenTransferRoutes,
   BrooderRoutes,
-  IncubatorRoutes
+  IncubatorRoutes,
+  ChickenHealthRoutes
 };
