+++ views/chicken-health-record.ejs
@@ -78,7 +78,7 @@
           </a>
         </li>
         <li class="nav-item">
-          <a class="nav-link text-white active bg-gradient-warning" href="/chicken-health-record">
+          <a class="nav-link text-white active bg-gradient-warning" href="/chicken-health/view">
             <div class="text-white text-center me-2 d-flex align-items-center justify-content-center">
               <i class="material-icons opacity-10">receipt_long</i>
             </div>
@@ -143,7 +143,7 @@
 
     <div class="row ms-2 mt-2 mb-0">
       <div class="col-4">
-        <button type="button" class="btn btn-sm btn-warning p-4 mb-0" onclick="location.href='/create-chicken-health-record'">
+        <button type="button" class="btn btn-sm btn-warning p-4 mb-0" onclick="location.href='/chicken-health/create'">
           Create new record
         </button>
       </div>
