+++ views/chicken-transfer.ejs
@@ -78,7 +78,7 @@
           </a>
         </li>
         <li class="nav-item">
-          <a class="nav-link text-white " href="/chicken-health-record">
+          <a class="nav-link text-white " href="/chicken-health/view">
             <div class="text-white text-center me-2 d-flex align-items-center justify-content-center">
               <i class="material-icons opacity-10">receipt_long</i>
             </div>
@@ -179,7 +179,7 @@
               </span>Total</p>
           </div>
           <div class="row mx-3">
-            <a class="btn btn-warning" href="/chicken/create?id=<%= data.coopID %>" role="button">Transfer</a>
+            <a class="btn btn-warning" href="/chicken-transfer/create?id=<%= data.coopID %>" role="button">Transfer</a>
           </div>
         </div>
       </div>
