+++ views/create-chicken-health-record.ejs
@@ -78,7 +78,7 @@
           </a>
         </li>
         <li class="nav-item">
-          <a class="nav-link text-white active bg-gradient-warning" href="/chicken-health-record">
+          <a class="nav-link text-white active bg-gradient-warning" href="/chicken-health/view">
             <div class="text-white text-center me-2 d-flex align-items-center justify-content-center">
               <i class="material-icons opacity-10">receipt_long</i>
             </div>
@@ -133,7 +133,7 @@
         <nav aria-label="breadcrumb">
           <ol class="breadcrumb bg-transparent mb-0 pb-0 pt-1 px-0 me-sm-6 me-5">
             <li class="breadcrumb-item text-sm"><a class="opacity-5 text-dark" href="/home">Dashboard</a></li>
-            <li class="breadcrumb-item text-sm"><a class="opacity-5 text-dark" href="/chicken-health-record">Chicken Health</a></li>
+            <li class="breadcrumb-item text-sm"><a class="opacity-5 text-dark" href="/chicken-health/view">Chicken Health</a></li>
             <li class="breadcrumb-item text-sm text-dark active" aria-current="page">Chicken Health Record</li>
           </ol>
           <h4 class="font-weight-bolder mb-0">Chicken Health Record</h4>
@@ -143,7 +143,7 @@
     <!------------------------------- Top Navbar ----------------------->
 
       <div class="p-3 mx-5 border border-secondary rounded shadow-lg mb-4">
-        <form class="needs-validation" action="/submit-chicken-health-record" method="post" role="form"
+        <form class="needs-validation" action="/chicken-health/submit" method="post" role="form"
           class="text-start">
           <div id="formRepeat">
             <div class="row border px-2 py-2 mx-1">
