+++ views/create-chicken-transfer-record.ejs
@@ -78,7 +78,7 @@
           </a>
         </li>
         <li class="nav-item">
-          <a class="nav-link text-white " href="/chicken-health-record">
+          <a class="nav-link text-white " href="/chicken-health/view">
             <div class="text-white text-center me-2 d-flex align-items-center justify-content-center">
               <i class="material-icons opacity-10">receipt_long</i>
             </div>
@@ -134,7 +134,7 @@
           <ol class="breadcrumb bg-transparent mb-0 pb-0 pt-1 px-0 me-sm-6 me-5">
             <li class="breadcrumb-item text-sm"><a class="opacity-5 text-dark" href="/home">Dashboard</a></li>
             <li class="breadcrumb-item text-sm"><a class="opacity-5 text-dark" href="/daily-record">Daily Record</a>
-            <li class="breadcrumb-item text-sm"><a class="opacity-5 text-dark" href="/chicken/view">Chicken</a>
+            <li class="breadcrumb-item text-sm"><a class="opacity-5 text-dark" href="/chicken-transfer/view">Chicken</a>
             </li>
             <li class="breadcrumb-item text-sm text-dark active" aria-current="page">Chicken Transfer Record</li>
           </ol>
@@ -152,7 +152,7 @@
     %>
 
       <div class="p-3 mx-5 border border-secondary rounded shadow-lg mb-4">
-        <form class="needs-validation" action="/chicken/submit" method="post" role="form" class="text-start">
+        <form class="needs-validation" action="/chicken-transfer/submit" method="post" role="form" class="text-start">
           <div id="formRepeat">
             <div class="row border px-2 py-2 mx-1">
               <div class="row mt-1">
