+++ views/daily-record.ejs
@@ -75,7 +75,7 @@
           </a>
         </li>
         <li class="nav-item">
-          <a class="nav-link text-white " href="/chicken-health-record">
+          <a class="nav-link text-white " href="/chicken-health/view">
             <div class="text-white text-center me-2 d-flex align-items-center justify-content-center">
               <i class="material-icons opacity-10">receipt_long</i>
             </div>
@@ -142,7 +142,7 @@
             <a class="btn btn-warning btn-lg" href="/coop/view" role="button">COOP</a>
           </div>
           <div class="col border py-6">
-            <a class="btn btn-warning btn-lg" href="/chicken/view" role="button">CHICKEN</a>
+            <a class="btn btn-warning btn-lg" href="/chicken-transfer/view" role="button">CHICKEN</a>
           </div>
         </div>
         <div class="row border">
