+++ views/incubator-record.ejs
@@ -112,7 +112,7 @@
           </a>
         </li>
         <li class="nav-item">
-          <a class="nav-link text-white " href="/chicken-health-record">
+          <a class="nav-link text-white " href="/chicken-health/view">
             <div class="text-white text-center me-2 d-flex align-items-center justify-content-center">
               <i class="material-icons opacity-10">receipt_long</i>
             </div>
